Renzo

+++ src/Web/Controllers/ProductController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Security.Claims;
 
+
 namespace Web.Controllers
 {
     [Route("api/[controller]")]
@@ -20,7 +21,7 @@ namespace Web.Controllers
         {
             _productService = productService;
         }
-        
+
         [HttpGet]
 
         public IActionResult GetAll()
@@ -32,29 +33,56 @@ namespace Web.Controllers
         }
 
         [HttpPost]
+        [Authorize]
 
         public IActionResult AddProduct([FromBody] ProductCreateRequest product)
 
         {
-            var newProduct = _productService.AddNewProduct(product);
-            return Ok(newProduct);
+            var userTypeString = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userTypeString == "Admin")
+            {
+                var newProduct = _productService.AddNewProduct(product);
+                return Ok(newProduct);
+            }
+            else
+            {
+                return Forbid();
+            }
         }
 
         [HttpDelete]
+        [Authorize]
 
         public IActionResult DeleteProduct([FromBody] int productId)
         {
+            var userTypeString = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userTypeString == "Admin")
+            {
 
-            _productService.DeleteProduct(productId);
+                _productService.DeleteProduct(productId);
             return Ok(new { message = "Product deleted successfully." });
+            }
+            else
+            {
+                return Forbid();
+            }
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public IActionResult UpdateProduct(int id, [FromQuery] string description, [FromQuery] double price, [FromQuery] int stock)
-        {          
-            _productService.UpdateProduct(id, description, price, stock);
-            return Ok(new { message = "Producto actualizado exitosamente." });
-                
+        {
+            var userTypeString = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userTypeString == "Admin")
+            {
+                _productService.UpdateProduct(id, description, price, stock);
+                return Ok(new { message = "Producto actualizado exitosamente." });
+            }
+            else
+            {
+                return Forbid();
+            }
+
         }
 
     }
